+++ airflow-core/src/airflow/api_fastapi/core_api/routes/public/dag_report.py
@@ -34,7 +34,7 @@ from airflow.api_fastapi.core_api.security import (
     ReadableDagsFilterDep,
     requires_access_dag,
 )
-from airflow.models.dagbag import DagBag
+from airflow.dag_processing.dagbag import DagBag
 
 dag_report_router = AirflowRouter(tags=["DagReport"], prefix="/dagReports")
