+++ airflow-core/src/airflow/models/__init__.py
@@ -98,7 +98,7 @@ __lazy_imports = {
     "BaseOperatorLink": "airflow.sdk",
     "BaseXCom": "airflow.sdk.bases.xcom",
     "Connection": "airflow.models.connection",
-    "DagBag": "airflow.models.dagbag",
+    "DagBag": "airflow.dag_processing.dagbag",
     "DagModel": "airflow.models.dag",
     "DagRun": "airflow.models.dagrun",
     "DagTag": "airflow.models.dag",
@@ -124,6 +124,7 @@ __lazy_imports = {
 if TYPE_CHECKING:
     # I was unable to get mypy to respect a airflow/models/__init__.pyi, so
     # having to resort back to this hacky method
+    from airflow.dag_processing.dagbag import DagBag
     from airflow.models.base import ID_LEN, Base
     from airflow.models.connection import Connection
     from airflow.models.dag import DagModel, DagTag
