+++ airflow-core/tests/unit/models/test_dagbag.py
@@ -35,10 +35,10 @@ import pytest
 from sqlalchemy import select
 
 from airflow import settings
+from airflow.dag_processing.dagbag import DagBag, _capture_with_reraise, _validate_executor_fields
 from airflow.exceptions import UnknownExecutorException
 from airflow.executors.executor_loader import ExecutorLoader
 from airflow.models.dag import DagModel
-from airflow.models.dagbag import DagBag, _capture_with_reraise, _validate_executor_fields
 from airflow.models.dagwarning import DagWarning, DagWarningType
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.sdk import DAG, BaseOperator
