+++ airflow-core/tests/unit/serialization/test_dag_serialization.py
@@ -46,6 +46,8 @@ from kubernetes.client import models as k8s
 
 import airflow
 from airflow._shared.timezones import timezone
+from airflow.config_templates.airflow_local_settings import DEFAULT_LOGGING_CONFIG
+from airflow.dag_processing.dagbag import DagBag
 from airflow.exceptions import (
     AirflowException,
     ParamValidationError,
